# core: don't include_str extension js code