--- a/core/runtime.rs
+++ b/core/runtime.rs
@@ -385,8 +385,9 @@ impl JsRuntime {
     for m in extensions.iter_mut() {
       let js_files = m.init_js();
       for (filename, source) in js_files {
+        let source = source()?;
         // TODO(@AaronO): use JsRuntime::execute_static() here to move src off heap
-        self.execute(filename, source)?;
+        self.execute(filename, &source)?;
       }
     }
     // Restore extensions