--- a/extensions/url/benches/url_ops.rs
+++ b/extensions/url/benches/url_ops.rs
@@ -10,7 +10,9 @@ fn setup() -> Vec<Extension> {
     Extension::builder()
       .js(vec![(
         "setup",
-        "const { URL } = globalThis.__bootstrap.url;",
+        Box::new(|| {
+          Ok(r#"const { URL } = globalThis.__bootstrap.url;"#.to_owned())
+        }),
       )])
       .build(),
   ]