This speeds up incremental rebuild when only touching JS files by 13-15%

Rebuild time after `touch 01_broadcast_channel.js`:

main: run 1 49.18s, run 2 50.34s
this: run 1 43.12s, run 2 43.19s

The reason for this speedup is that `deno_runtime(build.rs)` does not
need to be rebuilt on every js file change now.

Closes #10786