@lucacasonato 2021-05-29
Binary size reduction is around 0.9%.

latest canary: 81315944 bytes
this branch: 80632656 bytes